don't retrieve the secret before purging (#13614)
we've seen in the wild a situation where a keystore had a single invalid secret that couldn't be removed since that action did a retrieve before purging. The retrieve would fail causing the purge to never happen.
This commit skips the retrieval but checks if the secret exists to preserve the scenario of reporting an error when removing a secret that doesn't exist.

(cherry picked from commit 49303c3fc12b91bb6c0a354cf285aba9514c7d34)