make sure versions.yml is only read once into gradle.ext (#17865)
this removes the need to re-read the yaml file in so many places, and also avoids having to state the jruby version an extra time in the rubyUtils.gradle

It uses gradle.ext, computed in settings.gradle, as a place to store the versions. This is important to load the versions before any other buildscript {} block runs.